% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonym.R
\name{anonym}
\alias{anonym}
\title{Function to anonymize datasets so to create MWEs}
\usage{
anonym(df)
}
\arguments{
\item{df}{= dataframe}
}
\description{
Is a function that easily anonymizes the dataset so to remove personal information identification. It is useful for the creation of minimal working examples allowing the researcher to to discuss the statistical findings with other colleagues without giving away sensitive information. Its advantage over the package anonimizer() is that it does a better job in keeping the variance and the relation between factors in the data.
}
\details{
Source: https://www.jaredknowles.com/journal/2013/5/27/writing-a-minimal-working-example-mwe-in-r
}
