% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypestoordinals.R
\name{genotypesToOrdinals}
\alias{genotypesToOrdinals}
\title{Definition: Utility that takes variables containing minor and heteropositions for genotypes and convert these to a single ordinal variable. The new variable will contain zero if none of the alleles are present, one for heterozigous and two for homozigous.}
\usage{
genotypesToOrdinals(variables, data)
}
\description{
Definition: Utility that takes variables containing minor and heteropositions for genotypes and convert these to a single ordinal variable. The new variable will contain zero if none of the alleles are present, one for heterozigous and two for homozigous.
}
