% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfiles.R
\name{loadFiles}
\alias{loadFiles}
\title{A function to load required files in csv format}
\usage{
loadFiles(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Definition: This function is very specific for the data, but automates the
process of going into the /input directory (where all the stats files lies)
and import them into dataframes. In order for it to work one must have set the working
directory and added an /input subdirectory.
}
\examples{
df <- loadFiles ()
}
\keyword{csv}
\keyword{files}
\keyword{plain}
\keyword{text}
