% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadfreesurfer.R
\name{loadFreeSurfer}
\alias{loadFreeSurfer}
\title{Function used to load freesurfer data.}
\usage{
loadFreeSurfer(x)
}
\arguments{
\item{x}{= directory}
}
\description{
The segmentations must be named aseg.allsegs, it will parse
all files with .table extension. It can manage segmentations
and parcellations.
}
