% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-trailing.R
\name{trim.trailing}
\alias{trim.trailing}
\title{Function: trim.trailing(variable_name)}
\usage{
trim.trailing(x)
}
\arguments{
\item{x}{= variable}
}
\description{
Definition: Sometimes when files from SPSS are imported as dataframes to R,
or when one works with windows text files in linux some trailing whitespaces
creeps in. This function returns the same variable_name but without the
extra whitespaces. It only works for final whitespaces.
}
\details{
Example: df$variable <- trim.trailing (otherdf$variable)

Source: http://stackoverflow.com/a/2261149/4417072
}
