% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-descriptive-stats.R
\name{write_descriptive_stats}
\alias{write_descriptive_stats}
\title{Function that appends descriptive stats with describe()}
\usage{
write_descriptive_stats(vars, filename)
}
\arguments{
\item{vars}{= variables to be described}

\item{filename}{= directory path}
}
\description{
Useful in cases where one is working with a large number of
variables and needs to perform that several times to print
a report.
Originally from a question on SO: http://stackoverflow.com/a/43490045/4417072
}
