% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonym.R
\name{anonym}
\alias{anonym}
\title{Function to anonymize datasets so to create MWEs}
\usage{
anonym(df)
}
\arguments{
\item{df}{= dataframe}
}
\description{
This function takes a dataframe and creates a dummy dataset, but maintaining
sintactical relationships between the variables. This has been tested with
multivariate data and works.
}
\details{
Source: https://www.jaredknowles.com/journal/2013/5/27/writing-a-minimal-working-example-mwe-in-r
}
