% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Function to load dependencies}
\usage{
dependencies(dep)
}
\arguments{
\item{Conjunction}{of packages to check}
}
\description{
This function check for dependencies, if installed they are left as is,
otherwise they are installed
}
\keyword{installation}
\keyword{packages}
