% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-zero-cols.R
\name{remove_zero_cols}
\alias{remove_zero_cols}
\title{Description: A function to remove empty variables that might have ended up in the dataset after concatenating data.
https://stackoverflow.com/questions/21530168/remove-columns-with-zero-values-from-a-dataframe}
\usage{
remove_zero_cols(df)
}
\arguments{
\item{df}{= dataframe}
}
\description{
Description: A function to remove empty variables that might have ended up in the dataset after concatenating data.
https://stackoverflow.com/questions/21530168/remove-columns-with-zero-values-from-a-dataframe
}
