% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-divide-thousand.R
\name{transformDivideThousand}
\alias{transformDivideThousand}
\title{Function to perform transformation}
\usage{
transformDivideThousand(data_frame, listofvars)
}
\arguments{
\item{dataframe}{= dataframe}

\item{optional}{= vectorListOfVariables}
}
\description{
This function applies a transformation, dividing variables by
1000. If the vector is passed it applies the transformation to all variables
in the dataframe.
Source: http://stackoverflow.com/a/36912017/4417072
}
